requirements.txt
litellm
python-dotenv
edge_tts
pygame
SpeechRecognition
PyAudio
pydub
sounddevice
colorama
